Reformatting and reorganizing sharedId submodule
